requirements.txt
openai
chromadb
python-dotenv
ipykernel
typer
